restarted dev fixed bugs | added confirm on bulk delete

+++ components/data-table.tsx
@@ -22,6 +22,7 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table"
+import { useConfirm } from "@/hooks/use-confirm"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Trash } from "lucide-react"
@@ -41,6 +42,11 @@ export function DataTable<TData, TValue>({
     onDelete,
     disabled,
 }: DataTableProps<TData, TValue>) {
+    const [ConfirmationDialog, confirm] = useConfirm(
+        "Are you sure?",
+        "You are about to perform a bulk delete."
+    );
+
     const [sorting, setSorting] = React.useState<SortingState>([])
     const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
     []
@@ -66,6 +72,7 @@ export function DataTable<TData, TValue>({
 
     return (
         <div>
+            <ConfirmationDialog />
             <div className="flex items-center py-4">
         <Input
           placeholder={`Filter ${filterKey}...`}
@@ -81,9 +88,13 @@ export function DataTable<TData, TValue>({
                         size="sm"
                         variant="outline"
                         className="ml-auto font-normal text-xs"
-                        onClick={() => {
-                            onDelete(table.getFilteredSelectedRowModel().rows)
-                            table.resetRowSelection();
+                        onClick={async () => {
+                            const ok = await confirm();
+
+                            if (ok) {
+                                onDelete(table.getFilteredSelectedRowModel().rows)
+                                table.resetRowSelection();
+                            }
                         }}
                     >
                         <Trash className="size-4 mr-2" />
